Re-enable MSVC warning 4800 for compiling with chromium_code
Mainly change the code to avoid the warnings; in a few cases we have
to use explicit casts.

BUG=pdfium:29
R=tsepez@chromium.org

Review URL: https://codereview.chromium.org/1783023002 .

+++ core/src/fpdfapi/fpdf_parser/cpdf_syntax_parser.cpp
@@ -772,7 +772,8 @@ CPDF_Stream* CPDF_SyntaxParser::ReadStream(CPDF_Dictionary* pDict,
   GetNextWordInternal(nullptr);
 
   int numMarkers = ReadEOLMarkers(m_Pos);
-  if (m_WordSize == kEndObjStr.GetLength() && numMarkers != 0 &&
+  if (m_WordSize == static_cast<unsigned int>(kEndObjStr.GetLength()) &&
+      numMarkers != 0 &&
       FXSYS_memcmp(m_WordBuffer, kEndObjStr.GetPtr(), kEndObjStr.GetLength()) ==
           0) {
     m_Pos = streamStartPos;
